Rails
briefly review how our current application works.
1. The user navigates to our application. In our case, we do that using
a
local URL such as http://localhost:3000/say/hello.
2. Rails then matches the route pattern, which it previously split into
two
parts and analyzed.
The say part is taken to be the name of a controller, so Rails creates
a new instance of the Ruby class SayController (which it finds in
app/controllers/ say_controller.rb).
3. The next part of the pattern, hello, identifies an action. Rails
invokes a method of that name in the controller. This action method
creates a new Time object holding the current time and tucks it away in
the @time instance variable.
4. Rails looks for a template to display the result. It searches the
directory app/views for a subdirectory with the same name as the
controller (say) and in that subdirectory for a file named after the
action (hello.html.erb).
5. Rails processes this file through the ERB templating system,
executing any embedded Ruby and substituting in values set up by the
controller.
6. The result is returned to the browser, and Rails finishes processing
this request.